lab1/requirements.txt
xarm-python-sdk
numpy
